refactor(liblili2json): move get_module_summary
