+++ workspaces/liblili2json/liblili2json/__init__.py
@@ -3,55 +3,13 @@ Export LibreLingo courses in the JSON format expected by the web app
 """
 
 import itertools
-from slugify import slugify
 from .types import *
 from .utils import *
 from .dictionary import *
 from .challenge_types import *
+from .module import *
 
 __version__ = '0.1.0'
-
-
-def calculate_number_of_levels(nwords, nphrases):
-    """
-    Calculates how many levels a skill should have
-    """
-    return round(1 + (nwords / 7) + (nphrases / 5))
-
-
-def get_module_summary(module):
-    """
-    Get a module summary for the course meta data
-    """
-
-    def get_imageset(images):
-        if images and len(images) == 3 and all(images):
-            return {"imageSet": images}
-        else:
-            return {}
-
-    def get_summary(words, phrases):
-        words = [word.in_target_language[0] for word in words]
-        phrases = [phrase.in_target_language[0] for phrase in phrases]
-
-        return words + phrases
-
-    def get_levels(words, phrases):
-        return calculate_number_of_levels(len(words), len(phrases))
-
-    return {
-        "title": module.title,
-        "skills": [
-            {
-                **(get_imageset(skill.image_set)),
-                "summary": get_summary(skill.words, skill.phrases),
-                "practiceHref": slugify(skill.name),
-                "id": get_opaque_id(skill, "Skill"),
-                "title": skill.name,
-                "levels": get_levels(skill.words, skill.phrases),
-            } for skill in module.skills
-        ]
-    }
 
 
 def get_course_data(course):
