+++ workspaces/liblili2json/liblili2json/module.py
@@ -1,0 +1,37 @@
+from slugify import slugify
+from .utils import get_opaque_id, calculate_number_of_levels
+
+
+def get_module_summary(module):
+    """
+    Get a module summary for the course meta data
+    """
+
+    def get_imageset(images):
+        if images and len(images) == 3 and all(images):
+            return {"imageSet": images}
+        else:
+            return {}
+
+    def get_summary(words, phrases):
+        words = [word.in_target_language[0] for word in words]
+        phrases = [phrase.in_target_language[0] for phrase in phrases]
+
+        return words + phrases
+
+    def get_levels(words, phrases):
+        return calculate_number_of_levels(len(words), len(phrases))
+
+    return {
+        "title": module.title,
+        "skills": [
+            {
+                **(get_imageset(skill.image_set)),
+                "summary": get_summary(skill.words, skill.phrases),
+                "practiceHref": slugify(skill.name),
+                "id": get_opaque_id(skill, "Skill"),
+                "title": skill.name,
+                "levels": get_levels(skill.words, skill.phrases),
+            } for skill in module.skills
+        ]
+    }
